Stage through a module of its own, not through the filesystem backend
`openTempFile` served both backends but lived in `local.ts`, so the bucket
backend imported the filesystem backend — right code, wrong layering, and the
kind of thing a reviewer has to stop and re-derive. Pure move; behaviour
unchanged, and the local cache suite still covers it end to end.
