+++ server/services/previewCache/local.ts
@@ -6,11 +6,10 @@
 // backing up their volume is already backing this up — and losing it costs a
 // re-fetch, not data.
 
-import crypto from 'crypto';
-import fsSync from 'fs';
 import fs from 'fs/promises';
 import path from 'path';
 import { cachedBytes, coldestCached, forget } from '../../db/previewCache.js';
+import { openTempFile } from './tempFile.js';
 import type { LocalCacheConfig } from './config.js';
 
 /**
@@ -63,103 +62,6 @@ export interface LocalWriter {
   commit(): Promise<boolean>;
   /** Throw the partial file away. Safe to call twice. */
   abort(): Promise<void>;
-}
-
-/**
- * Bytes accumulating in a temp file, to be published or thrown away.
- *
- * ⚠ Extracted so the `s3` backend can stage through the same code rather than
- * collecting an object in memory to PUT it. Both backends have the same problem —
- * chunks arrive from a stream, the destination is not a heap buffer — and only
- * the publish step differs: `local` renames, `s3` uploads and unlinks.
- */
-export interface TempFile {
-  path: string;
-  write(chunk: Buffer): void;
-  /** Close, and report whether the bytes are intact. The file is the caller's to
-   *  move or delete afterwards; a `false` has already cleaned up. */
-  close(): Promise<boolean>;
-  /** Close and delete, whatever state it is in. Safe to call twice. */
-  discard(): Promise<void>;
-}
-
-export async function openTempFile(dir: string, name: string): Promise<TempFile | null> {
-  // ⚠⚠ RANDOM, not pid+timestamp. Nothing dedupes the byte path — `mediaPool`
-  // bounds concurrency, not identity — so several writes of the SAME key can be in
-  // flight at once. Two in the same millisecond shared a temp path, and a write
-  // truncates only at open: the shorter one left the longer body's tail in place
-  // and published a file matching NEITHER, under a row recording the shorter
-  // length. Reproduced with 4 MB and 1 MB bodies while this was pid+timestamp.
-  const tmp = path.join(dir, `${name}.${crypto.randomUUID()}.tmp`);
-  let handle: fsSync.WriteStream;
-  try {
-    await fs.mkdir(dir, { recursive: true });
-    handle = fsSync.createWriteStream(tmp);
-  } catch {
-    return null;
-  }
-  // A write error (ENOSPC, EROFS) must not become an unhandled 'error' event and
-  // take the process down; it is recorded and turns commit into a no-op.
-  let broken = false;
-  handle.on('error', () => {
-    broken = true;
-  });
-
-  let done = false;
-  /**
-   * ⚠⚠ Waits on `close`, NOT on `end`'s callback.
-   *
-   * `writable.end(cb)` attaches its callback to `finish`, which a stream that
-   * ERRORS never reaches — node's own docs say the callback "may or may not" be
-   * called on error. So on ENOSPC or EROFS, which is precisely the state a cache
-   * ceiling exists for, the promise never settled: `commit()` and `abort()` hung
-   * forever, the descriptor and the temp file stayed stranded, and the store
-   * decision was never reached. `close` fires on both paths.
-   *
-   * ⚠ The already-closed check is not belt-and-braces. A stream destroyed by an
-   * earlier error has emitted `close` before we ever get here, and a listener
-   * added afterwards waits for an event that has been and gone.
-   */
-  const closeStream = () =>
-    new Promise<void>((resolve) => {
-      if (handle.closed || handle.destroyed) {
-        resolve();
-        return;
-      }
-      handle.once('close', () => resolve());
-      handle.end();
-    });
-
-  return {
-    path: tmp,
-    write(chunk: Buffer): void {
-      if (done || broken) return;
-      // Backpressure is deliberately not awaited: the response is already being
-      // piped to the reader and must not wait on our disk. The stream's own buffer
-      // bounds this at the transfer's cap, which is what the memory fix is about.
-      handle.write(chunk);
-    },
-    async close(): Promise<boolean> {
-      if (done) return false;
-      done = true;
-      await closeStream();
-      if (broken) {
-        await fs.unlink(tmp).catch(() => {});
-        return false;
-      }
-      return true;
-    },
-    async discard(): Promise<void> {
-      if (!done) {
-        done = true;
-        await closeStream();
-      }
-      // ⚠ The temp file is OURS. Left behind it is invisible to the index and
-      // therefore to eviction — bytes on the volume nothing counts and nothing can
-      // reclaim.
-      await fs.unlink(tmp).catch(() => {});
-    },
-  };
 }
 
 export async function openLocalWrite(
